Add s2s-utils
