+++ package.json
@@ -12,8 +12,6 @@
     "fmt": "prettier --write '**/*.{js,json}'",
     "lint": "eslint .",
     "prebuild": "rimraf ./packages/**/lib",
-    "prepublish": "yarn build",
-    "publish": "lerna publish",
     "build": "lerna run build",
     "prepare": "yarn build",
     "postinstall": "yarn fmt"
