+++ packages/babel-plugin-s2s-action-root/package.json
@@ -24,8 +24,8 @@
     "babel-template": "^6.26.0",
     "babel-types": "^6.26.0",
     "globby": "^6.1.0",
+    "s2s-utils": "^0.0.0",
     "s2s-babel-flow-types": "^0.1.0",
-    "slash": "^1.0.0",
     "uppercamelcase": "^3.0.0"
   }
 }
