+++ packages/babel-plugin-s2s-state-root/package.json
@@ -25,7 +25,7 @@
     "babel-types": "^6.26.0",
     "globby": "^6.1.0",
     "s2s-babel-flow-types": "^0.1.0",
-    "slash": "^1.0.0",
+    "s2s-utils": "^0.0.0",
     "uppercamelcase": "^3.0.0"
   }
 }
