+++ packages/s2s-utils/.babelrc
@@ -1,0 +1,3 @@
+{
+  "presets": ["env"]
+}
