+++ packages/s2s-utils/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "s2s-utils",
+  "version": "0.0.0",
+  "description": "s2s utils",
+  "license": "MIT",
+  "repository": "akameco/s2s-plugins",
+  "author": "akameco <akameco.t@gmail.com>",
+  "main": "lib/index.js",
+  "files": ["lib"],
+  "scripts": {
+    "prepare": "npm run build",
+    "build": "babel src --ignore *.test.js -d lib "
+  },
+  "dependencies": {
+    "slash": "^1.0.0"
+  }
+}
