+++ packages/s2s-utils/src/index.test.js
@@ -1,0 +1,17 @@
+// @flow
+import { getImportPath } from '.'
+
+test('getImportPath same folder', () => {
+  const result = getImportPath('path/to/index.js', 'path/to/test.js')
+  expect(result).toBe('./test')
+})
+
+test('getImportPath same folder when not ext', () => {
+  const result = getImportPath('path/to/index', 'path/to/test')
+  expect(result).toBe('./test')
+})
+
+test('getImportPath parent folder', () => {
+  const result = getImportPath('path/to/nest/index.js', 'path/to/test.js')
+  expect(result).toBe('../test')
+})
